requirements.txt
streamlit
python-dotenv
google-generativeai
faiss-cpu
sentence-transformers
